requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
coverage==6.3.1
defusedxml==0.7.1
Django==4.0.2
django-stubs==1.9.0
django-stubs-ext==0.3.1
djangorestframework==3.13.1
docformatter==1.4
environ==1.0
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
types-pytz==2021.3.5
types-PyYAML==6.0.4
typing-extensions==4.1.1
untokenize==0.1.1
urllib3==1.26.8
